pyproject.toml
[project]
name = "qwen"
version = "0.1.0"
description = "Process images using Qwen2-VL-7B model with optional resizing and custom prompt."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio-client>=1.3.0",
    "pillow>=10.4.0",
]
